Refactor `SignMessageRequestContentView` out of `SignMessageRequestView` for re-use. Update `SaferSignMessageRequestContainerView` to use `SignMessageRequestContentView` for details view display.
