+++ Sources/BraveWallet/Panels/Signature Request/SignMessageRequestContainerView.swift
@@ -52,6 +52,8 @@ struct SignMessageRequestContainerView: View {
           ),
           cowSwapOrder: cowSwapOrder,
           ethSwapDetails: store.ethSwapDetails[store.currentRequest.id],
+          needPilcrowFormatted: $store.needPilcrowFormatted,
+          showOrignalMessage: $store.showOrignalMessage,
           nextTapped: store.next,
           action: handleAction(approved:)
         )
